requirements.txt
google-auth>=2.26.1
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
gspread>=5.12.0
python-dotenv>=1.0.1
openai>=1.60.0
requests>=2.32.3
pandas>=2.1.3
validators>=0.22.0
beautifulsoup4>=4.12.2
urllib3>=2.0.7
pydantic>=2.10.4,<2.11.0 